# server/doc/openapi.yaml
openapi: 3.0.3
info:
  title: Hub API
  description: APIs for dashboard
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /export-data:
    get:
      tags:
        - Dashboard
      summary: Export data within a date range
      operationId: exportData
      parameters:
        - $ref: "#/components/parameters/dateFrom"
        - $ref: "#/components/parameters/dateTo"
      responses:
        "200":
          description: CSV file containing the requested metrics
          content:
            text/csv:
              schema:
                type: string
                format: binary
...
components:
...
  parameters:
...
    dateFrom:
      in: query
      schema:
        type: string
        format: date-time
      required: false
      name: dateFrom
    dateTo:
      in: query
      schema:
        type: string
        format: date-time
      required: false
      name: dateTo